Extracts psql query run to its own class to be reusable
