+++ lib/cartodb/api/query_tables_api.js
@@ -1,9 +1,5 @@
-var PSQL = require('cartodb-psql');
-var step = require('step');
-
-function QueryTablesApi(pgConnection, metadataBackend) {
-    this.pgConnection = pgConnection;
-    this.metadataBackend = metadataBackend;
+function QueryTablesApi(pgQueryRunner) {
+    this.pgQueryRunner = pgQueryRunner;
 }
 
 var affectedTableRegexCache = {
@@ -20,7 +16,7 @@ QueryTablesApi.prototype.getAffectedTablesInQuery = function (username, sql, cal
 
     var query = 'SELECT CDB_QueryTables($windshaft$' + prepareSql(sql) + '$windshaft$)';
 
-    this.runQuery(username, query, handleAffectedTablesInQueryRows, callback);
+    this.pgQueryRunner.run(username, query, handleAffectedTablesInQueryRows, callback);
 };
 
 function handleAffectedTablesInQueryRows(err, rows, callback) {
@@ -46,7 +42,7 @@ QueryTablesApi.prototype.getAffectedTablesAndLastUpdatedTime = function (usernam
         'WHERE m.tabname = any ((SELECT tablenames from querytables)::regclass[])'
     ].join(' ');
 
-    this.runQuery(username, query, handleAffectedTablesAndLastUpdatedTimeRows, callback);
+    this.pgQueryRunner.run(username, query, handleAffectedTablesAndLastUpdatedTimeRows, callback);
 };
 
 function handleAffectedTablesAndLastUpdatedTimeRows(err, rows, callback) {
@@ -68,43 +64,6 @@ function handleAffectedTablesAndLastUpdatedTimeRows(err, rows, callback) {
         lastUpdatedTime: lastUpdatedTime * 1000
     });
 }
-
-
-QueryTablesApi.prototype.runQuery = function(username, query, queryHandler, callback) {
-    var self = this;
-
-    var params = {};
-
-    step(
-        function setAuth() {
-            self.pgConnection.setDBAuth(username, params, this);
-        },
-        function setConn(err) {
-            if (err) {
-                throw err;
-            }
-            self.pgConnection.setDBConn(username, params, this);
-        },
-        function executeQuery(err) {
-            if (err) {
-                throw err;
-            }
-            var psql = new PSQL({
-                user: params.dbuser,
-                pass: params.dbpass,
-                host: params.dbhost,
-                port: params.dbport,
-                dbname: params.dbname
-            });
-            psql.query(query, function(err, resultSet) {
-                resultSet = resultSet || {};
-                var rows = resultSet.rows || [];
-                queryHandler(err, rows, callback);
-            });
-        }
-    );
-};
-
 
 function prepareSql(sql) {
     return sql
