+++ lib/cartodb/server_options.js
@@ -5,6 +5,7 @@ var assert = require('assert');
 var RedisPool = require('redis-mpool');
 
 var QueryTablesApi = require('./api/query_tables_api');
+var PgQueryRunner = require('./backends/pg_query_runner');
 var PgConnection = require('./backends/pg_connection');
 var TemplateMaps = require('./template_maps.js');
 var MapConfigNamedLayersAdapter = require('./models/mapconfig_named_layers_adapter');
@@ -25,11 +26,12 @@ var REQUEST_QUERY_PARAMS_WHITELIST = [
 module.exports = function(redisPool) {
     redisPool = redisPool || new RedisPool(_.extend(global.environment.redis, {name: 'windshaft:server_options'}));
 
-    var cartoData = require('cartodb-redis')({ pool: redisPool }),
-        lzmaWorker = new LZMA(),
-        pgConnection = new PgConnection(cartoData),
-        queryTablesApi = new QueryTablesApi(pgConnection, cartoData),
-        cdbRequest = new CdbRequest();
+    var cartoData = require('cartodb-redis')({ pool: redisPool });
+    var lzmaWorker = new LZMA();
+    var pgConnection = new PgConnection(cartoData);
+    var pgQueryRunner = new PgQueryRunner(pgConnection);
+    var queryTablesApi = new QueryTablesApi(pgQueryRunner);
+    var cdbRequest = new CdbRequest();
 
     var rendererConfig = _.defaults(global.environment.renderer || {}, {
         cache_ttl: 60000, // milliseconds
